refactor: compressor tests moved and fixed review tasks
